+++ src/main/java/com/ebs/repository/DatabaseProfileRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ebs.entity.DatabaseProfile;
 
-public interface DatabaseProfileRepository extends JpaRepository<DatabaseProfile,  Integer > {
+public interface DatabaseProfileRepository extends JpaRepository<DatabaseProfile,  Long > {
 
 	DatabaseProfile findByProfileName(String profileName);
 }
